Pinning mongoDB library at v2.0.8 for now
A changed was made recently to support the newer authentication
mechanisms in MongoDB 3.x. This change has broken our code in a way
that I don't yet understand. For now, holding back to v2.0.8 until
it can be investigated further. This may also require changes to
the mongodb server and or the dbs in it.