Changed KSDL monad so we can log more effectively
What used to be (ExceptT String) is now (ExceptT ErrMsg) in the KSDL
monad. This is so we can see the type of error that comes out of
the location process and have an effect on the log Priority that's
used. This is then being used with the new stderr log handler to make
some severe errors express to stderr and thereby notify an operator.