Put the code back to not use from/toBSON
This was proving to be problematic in the toBSON direction with
strange field values showing up in MongoDB like NumberLong(...).

I think the approach from ks-library to convert From/ToJSON-deriving
types to/from BSON through Aeson wasn't a good idea. This commit
is part of the way to undoing that.

What was also done here is removal of the reliance on MongoDB for
the regions list. This is now in a Map Text (Text, Text) and is
hard-coded for now. Eventually this will be in a config file.