Added new fields for regional_stats
The location and display_name fields are in support of the new
location-based API for maniuplating this data.

As for the display_name field, this will eventually replace the
state and county fields. state and county are now deprecated.

Also changed is the way we carry time zone info for a source in
source config files. These are now simple INTs reflecting the
number of hours offset from GMT (WITHOUT any daylight-savings
info!) This is simpler (and apparently less breaky) than the
serialized Data.Time.LocalTime.TimeZone data structure we were using.