Added distance calculations to Places matches
We are now computing the distance between each Places result and
the geocoding coordinates we get back. The idea is we may be able
to use this to enhance matches and fix some stubborn problem matches.

This commit also includes a big redesign of the
KS.Locate.Places.Match.match machinery. It's much less complicated
now.