Upgraded from Stackage lts-9.4 to lts-11.10
The purpose of this upgrade was to get a newer version of wreq with
updated tls security.

This upgrade required a change to get the (=~) function from regex-pcre
instead of regex-compat, from which is was removed. (Makes sense, the =~
operator is a Perl thing really)

Also, oddly, found a binary in the cabal file that was missing the
hcoord module, but mysteriously built fine until now!