Switched from nearbysearch to findplacefromtext
This is a newer, more accurate, and less expensive API. We think this
will get us more accurate matches in a lot of cases.

Some development scripts were also added that were used to hammer out
some good query strings.

Another change was made to automatically use the returned `types` field
from the Places API to reject all supermarkets and grocery stores. We're
better off not trying to list these types of establishments as the
various food counters inside to not map to a single Google Place record.